addphoto_Api

+++ app/src/main/java/com/example/foursquare/model/FavouritesDataModel.kt
@@ -5,7 +5,7 @@ import com.beust.klaxon.*
 private val klaxon = Klaxon()
 
 data class Favourites (
-    val status: Long,
+    val status: Int,
     val error: String,
     val message: String,
     val pageNo: Long,
